+++ share.sh
@@ -1,0 +1,4 @@
+src=$1
+dst=$2
+txnid=$3
+python3 src/run.py -f share -s $src -r $dst -t $txnid
