+++ src/constants.py
@@ -1,0 +1,5 @@
+from enum import Enum
+class TxnType(Enum):
+    UPLOAD = 1
+    SHARE = 2
+    INVALID = -1
